// client/src/types/QuizInfo.ts
export interface QuizInfo {
    quiz_id: number;
    quiz_link: string;
    number_of_questions: number;
    created_at: string;
    quiz_name: string;
    description: string;
    first_name: string;
    last_name: string;
    course_name: string;
    subject_name: string;
    attempt_count: number;
}